Review Controller Refactor
Refactor review controller to reduce complexity
