+++ src/utils/check.box.ticked.ts
@@ -1,0 +1,8 @@
+export const isStatementCheckboxTicked = (checkboxValue: string): boolean => {
+
+  if (checkboxValue === "true") {
+    return true;
+  }
+
+  return false;
+};
